% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessions_warning.R
\name{accessions_warning}
\alias{accessions_warning}
\title{Get acessions and organism for each protein identifier}
\usage{
accessions_warning(protein_ids)
}
\arguments{
\item{protein_ids}{A string vector containing protein identifiers.}
}
\description{
Core function used by \code{\link{getAccessions}}.
}
\author{
Jose V. Die
}
