% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archids_warning.R
\name{archids_warning}
\alias{archids_warning}
\title{Get architecture identifiers for the conserved domains}
\usage{
archids_warning(gene_family)
}
\arguments{
\item{gene_family}{A string with conserved domain(s).}
}
\description{
Parses SPARCLE database (NCBI) and extract electronic identifiers for 
each conserved domain.
}
\author{
Jose V. Die
}
