% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_proteins.R
\name{extract_proteins}
\alias{extract_proteins}
\title{Get the protein identifiers}
\usage{
extract_proteins(targets, taxonIds)
}
\arguments{
\item{targets}{A string with the electronic links for the SPARCLE architecture.}

\item{taxonIds}{A string with the taxonomic species identifiers;
legume species (by default).}
}
\description{
Extract the protein identifier for the given taxonomic species, which are
   hosted by the RefSeq database (NCBI).
}
\details{
First, get the protein ids from RefSeq database. Then, extract only the ids
  for the selected taxonomic species (by default, legume species).
}
\author{
Jose V. Die
}
