% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_XP_from_spp.R
\name{extract_XP_from_spp}
\alias{extract_XP_from_spp}
\title{Get the XP accession ids for a selected Species}
\usage{
extract_XP_from_spp(target, myspp)
}
\arguments{
\item{target}{A vector/list object with protein ids.}

\item{myspp}{A string with the species name to extract the XP ids.}
}
\value{
A string vector of XP ids for selected species.
}
\description{
Parse a string vector with the protein electronic ids and extract only
those for a selected species.
}
\details{
The species name requires Genus or Genus species (NCBI nomenclature).
For example, for extracting only the chickpea sequences from a protein ids, use
: "Cicer" or "Cicer arietinum".
}
\examples{
my_values <- c("571493944",  "828328898",  "1117487898", "1431691079",
 "828337470",  "502078116")
extract_XP_from_spp(my_values, "Cicer arietinum")

}
\author{
Jose V. Die

Kimberly H. LeBlanc
}
