% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterarchids_warning.R
\name{filterarchids_warning}
\alias{filterarchids_warning}
\title{Filter protein architectures based on conserved domains}
\usage{
filterarchids_warning(archs_ids, filter)
}
\arguments{
\item{archs_ids}{A string with the architecture identifiers.}

\item{filter}{A string with the domains as filter.}
}
\description{
Parse the architecture identifiers and extract those that contain, at
least, the conserved domaind selected as filter.
}
\author{
Jose V. Die
}
