% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneHummus.R
\docType{package}
\name{geneHummus}
\alias{geneHummus}
\alias{geneHummus-package}
\title{genehummus: A pipeline to define gene families in Legumes and beyond}
\description{
genehummus is a pipeline with high specificity and sensitivity in
extracting proteins from the RefSeq database (NCBI).
}
\seealso{
see the preprint in \href{https://www.biorxiv.org/content/10.1101/436659v1}{BioRvix}
}
\author{
Jose V. Die \email{jose.die@uco.es}, Moamen M. Elmassry,
Kimberly H. LeBlanc, Olaitan I. Awe, Allissa Dillman, Ben Busby
}
