% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp.R
\name{get_spp}
\alias{get_spp}
\title{Get the species name from the description sequence}
\usage{
get_spp(description)
}
\arguments{
\item{description}{A string vector with the sequence description (title and species).}
}
\value{
for each sequence description, extract the species name.
}
\description{
Parse a string vector with sequence descriptions (title and species) and
extract the species name.
}
\author{
Jose V. Die
}
