% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProtlinks.R
\name{getProtlinks}
\alias{getProtlinks}
\title{Get the protein identifiers for a given architecture}
\usage{
getProtlinks(archs_ids)
}
\arguments{
\item{archs_ids}{A string with the architecture identifiers
(SPARCLE database, NCBI)}
}
\description{
Parse the RefSeq database and extract all the protein identifiers that
have a given architecture.
}
\author{
Jose V. Die
}
