% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSparcleArchs.R
\name{getSparcleArchs}
\alias{getSparcleArchs}
\title{Get the electronic architecture for a conserved domain}
\usage{
getSparcleArchs(CD)
}
\arguments{
\item{CD}{A string with the conserved domain(s)}
}
\description{
Parses the SPARCLE database (NCBI) and extract the electronic links for
a given conserved domain.
}
\examples{
\dontrun{
CD = "pfam02362"
getSparcleArchs(CD) }

}
\author{
Jose V. Die
}
