% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_warning.R
\name{labels_warning}
\alias{labels_warning}
\title{Get description label for a protein architecture identifier}
\usage{
labels_warning(arch_ids)
}
\arguments{
\item{arch_ids}{A string with the architecture electronic identifiers.}
}
\description{
Parses the architecture identifier and extract the corresponding labels.
}
\author{
Jose V. Die
}
