% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legumesIds.R
\docType{data}
\name{legumesIds}
\alias{legumesIds}
\title{NCBI taxonomy ids for the legume family}
\format{a numeric vector with 10.009 elements}
\source{
Taxonomy identifiers for \href{https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?id=3803}{Fabaceae}
  species (Taxonomy databse, NCBI).
}
\usage{
data(legumesIds)
}
\description{
Taxonomy identifier for about 10,000 legume species
}
\keyword{datasets}
