% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins_warning.R
\name{proteins_warning}
\alias{proteins_warning}
\title{Get RefSeq protein identifiers for the given taxonomic species}
\usage{
proteins_warning(arch_ids, taxonIds)
}
\arguments{
\item{arch_ids}{A string with the electronic links for the SPARCLE.}

\item{taxonIds}{A vector string with taxonomy ids.}
}
\description{
Parse the RefSeq database using protein architecture identifiers and 
extract protein ids. for selected taxonomic species. Core function used 
by \code{\link{getProteins_from_tax_ids}}.
}
\author{
Jose V. Die
}
