% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetIds.R
\docType{data}
\name{sizeIds}
\alias{sizeIds}
\title{Build a list containing N elements per element list}
\format{An object of class \code{numeric} of length 1.}
\usage{
sizeIds
}
\description{
Split a vector into N elements, so that each element contains a given
length.
}
\author{
Jose V. Die
}
\keyword{datasets}
