% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArch_labels.R
\name{getArch_labels}
\alias{getArch_labels}
\title{Get the description label for a protein architecture identifier}
\usage{
getArch_labels(arch_ids)
}
\arguments{
\item{arch_ids}{A string with the architecture electronic identifiers.}
}
\value{
print out the description label for the candidate architectures that
contain the proteins we are looking for.
}
\description{
Parses the architecture identifiers and extract their corresponding
labels.
}
\examples{
\dontshow{
filtered_archids <- c("12034188", "12034184","12034182")
getArch_labels(filtered_archids)}
\donttest{
filtered_archids <- filterArch_ids(archs_ids, my_filter)
getArch_labels(filtered_archids)}

}
\seealso{
\code{filterArch_ids}
}
\author{
Jose V. Die
}
